gst-nmos-rs: extract Side and UdpMedia into leaf type modules
Move Side to session/types and UdpMedia/UdpLeg to session/udp/types so
sdp and inner depend on shared data shapes without pulling in UDP
activation logic. Convert session/udp to a directory module.
